Added protected routing and fixed errors
