+++ app/package.json
@@ -31,17 +31,20 @@
 		"@shopify/polaris": "^12.21.0",
 		"@types/feathersjs__feathers": "^3.1.14",
 		"babel-plugin-import": "^1.13.8",
-		"envalid": "^5.0.0",
+		"copy-webpack-plugin": "^12.0.2",
+		"envalid": "^8.0.0",
 		"install": "^0.13.0",
-		"isomorphic-fetch": "^2.2.1",
+		"isomorphic-fetch": "^3.0.0",
 		"lodash": "^4.17.21",
 		"moment": "^2.30.1",
 		"npm": "^10.5.0",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
+		"react-router-dom": "^6.26.1",
 		"socket.io-client": "^2.3.0",
 		"terser-webpack-plugin": "^5.3.10",
-		"tsconfig-paths-webpack-plugin": "^4.1.0"
+		"tsconfig-paths-webpack-plugin": "^4.1.0",
+		"workbox-webpack-plugin": "^7.1.0"
 	},
 	"devDependencies": {
 		"@babel/core": "^7.24.0",
