+++ app/src/pages/dashboard.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Dashboard() {
+
+  return (
+    <h1>You have logged in!</h1>
+  )
+
+}
